requirements.txt
pymongo>=4.5.0
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.1.0
Flask>=3.0.0
Flask-CORS>=4.0.0
